src/requirements.txt
six==1.3.0
tornado==3.1
motor==0.1
sockjs-tornado==1.0.0
lxml==3.1.0
jsonschema==1.2.0
pyzmq==13.1
Momoko==1.0.0
toro==0.5